rename: component 이름 수정
